+++ Changelog.md
@@ -1,4 +1,10 @@
 # Changelog
+## Alpha 0.0.3
+Released on 14th March 2018
+* Added "swing" method for LivingEntity objects to allow them to swing their arms too, like Player objects
+* Reimplemented "jump" method for LivingEntity objects using some nasty reflection
+* Added an "explode" method for both Player and LivingEntity objects. Fairly self-explanatory. KABOOM!
+
 ## Alpha 0.0.2
 Released on 13th March 2018  
 * Added code comments. A double forward slash // will cause the rest of the line to be ignored when building Java files. Example: `tell player to jump //Causes the player to jump in the air!`